dags/hitex_amazon_fba_pipeline.py
"""
Amazon FBA Pipeline - Handles 10M row inventory reports
"""
from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.operators.dummy import DummyOperator

default_args = {
    'owner': 'hitex-data-engineering',
    'depends_on_past': False,
    'start_date': datetime(2024, 1, 1),
    'retries': 3,
    'retry_delay': timedelta(minutes=5)
}

def extract_amazon_listing_reports():
    """Extract 10K row listing reports"""
    # Implementation for Amazon Seller API
    pass

def extract_amazon_inventory_transactions():
    """Extract 10M row inventory transactions with chunking"""
    # Implementation for large dataset
    pass

with DAG(
    'hitex_amazon_fba_pipeline',
    default_args=default_args,
    description='Amazon FBA data pipeline - 10M row handling',
    schedule_interval='0 */2 * * *',  # Every 2 hours
    catchup=False,
    tags=['amazon', 'fba', 'inventory']
) as dag:
    
    start = DummyOperator(task_id='start')
    end = DummyOperator(task_id='end')
    
    extract_listing = PythonOperator(
        task_id='extract_listing_reports',
        python_callable=extract_amazon_listing_reports
    )
    
    extract_inventory = PythonOperator(
        task_id='extract_inventory_transactions',
        python_callable=extract_amazon_inventory_transactions
    )
    
    start >> [extract_listing, extract_inventory] >> end