requirements/requirements-test.txt
# Core Airflow Dependencies
apache-airflow==2.7.3
apache-airflow-providers-google==10.10.1
apache-airflow-providers-postgres==5.7.1

# Data Processing
pandas==2.1.4
numpy==1.24.3
pyarrow==14.0.1

# Data Quality & Testing
great-expectations==0.18.5
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-airflow==0.0.3

# Code Quality
black==23.11.0
pylint==3.0.3
flake8==6.1.0
isort==5.12.0

# BigQuery & GCP
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4

# Development & Testing
docker==6.1.3
testcontainers==3.7.1
faker==20.1.0